chore: Refactor journal scraping functionality into CLI and domain modules
- Moved journal scraping logic from procedures to a new CLI entry point in `philoch_bib_enhancer.cli.crossref_journal_scraping_cli`.
- Created domain logic for bibkey matching in `philoch_bib_enhancer.domain.bibkey_matching`.
- Updated imports to reflect new module structure, including changes from `procedures` to `ports` and `domain`.
- Introduced configuration models for environment variables and bibkey matching.
- Removed the old `journal_scraping.py` procedure file to streamline the codebase.
- Added a new entry point for the CLI in `pyproject.toml` for easier access to the journal scraping functionality.
