+++ philoch_bib_enhancer/cli/__init__.py
@@ -1,0 +1,1 @@
+"""CLI entry points - imperative shell that wires dependencies."""
