+++ philoch_bib_enhancer/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Domain logic module - pure functions with no I/O or side effects."""
